+++ file_options.go
@@ -12,42 +12,52 @@ import (
 
 // Options define the behavior of `WriteReaderAtomic()`, etc.
 type Options struct {
-	defaultFileMode os.FileMode
-	useFileMode     os.FileMode
-	keepFileMode    bool
-	flag            int
+	defaultFileMode os.FileMode // default 0
+	fileMode        os.FileMode // default 0
+	flags           int         // default 0
+	keepFileMode    bool        // default false
+	setSymlinkOwner bool        // default false
 }
 
 // Option functions modify Options.
 type Option func(*Options)
 
-// UseFileMode sets the file mode to the desired value and has precedence over all
-// other options.
-func UseFileMode(mode os.FileMode) Option {
-	return func(opts *Options) {
-		opts.useFileMode = mode
-	}
-}
-
-// DefaultFileMode sets the default file mode instead of using the
+// WithDefaultFileMode sets the default file mode instead of using the
 // `os.CreateTemp()` default of `0600`.
-func DefaultFileMode(mode os.FileMode) Option {
+func WithDefaultFileMode(mode os.FileMode) Option {
 	return func(opts *Options) {
 		opts.defaultFileMode = mode
 	}
 }
 
-// KeepFileMode preserves the file mode of an existing file instead of using the
+// WithFileMode sets the file mode to the desired value and has precedence over all
+// other options.
+func WithFileMode(mode os.FileMode) Option {
+	return func(opts *Options) {
+		opts.fileMode = mode
+	}
+}
+
+// WithFlags sets the flag option.
+func WithFlags(flags int) Option {
+	return func(opts *Options) {
+		opts.flags = flags
+	}
+}
+
+// WithKeepFileMode preserves the file mode of an existing file instead of using the
 // default value.
-func KeepFileMode(keep bool) Option {
+func WithKeepFileMode(keep bool) Option {
 	return func(opts *Options) {
 		opts.keepFileMode = keep
 	}
 }
 
-// Flag sets the flag option.
-func Flag(flag int) Option {
+// WithSetSymlinkOwner sets the symlink's owner to be the current user.
+// Otherwise, the symlink will have a default owner assigned by the system,
+// such as BUILTIN\Administrator.
+func WithSetSymlinkOwner(setSymlinkOwner bool) Option {
 	return func(opts *Options) {
-		opts.flag = flag
+		opts.setSymlinkOwner = setSymlinkOwner
 	}
 }